requirement.txt
Django==1.6.5
MySQL-python==1.2.5
Pillow==2.4.0
South==0.8.4
Werkzeug==0.9.6
argparse==1.2.1
django-appconf==0.6
django-compressor==1.4
django-extensions==1.3.8
djangorestframework==2.3.14
ipdb==0.8
ipython==2.1.0
six==1.7.2
wsgiref==0.1.2
django-categories==1.2.2
django-taggit==0.12
#photomanagements
git+https://github.com/siteshwar/django-photologue.git@algin#egg=django_photologue